+++ R/extract_points.R
@@ -2,7 +2,7 @@
 #'
 #' @description Extract points
 #'
-#' @param raster RasterLayer with environmental data
+#' @param raster SpatRaster with environmental data
 #' @param pattern ppp object with point pattern.
 #'
 #' @details
@@ -15,7 +15,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' landscape_classified <- classify_habitats(landscape, n = 5, style = "fisher")
+#' landscape_classified <- classify_habitats(terra::rast(landscape), n = 5, style = "fisher")
 #' extract_points(raster = landscape_classified, pattern = species_b)
 #' }
 #'
@@ -25,12 +25,12 @@
 #' @keywords internal
 extract_points <- function(raster, pattern){
 
-  habitat_levels <- sort(unique(raster::values(raster))) # get all habitats sorted
+  habitat_levels <- sort(unique(terra::values(raster))) # get all habitats sorted
 
   pattern <- spatstat.geom::coords(pattern) # extract only coords of points
 
   # get habitat points are located within
-  pattern_extracted <- factor(raster::extract(x = raster, y = pattern, factor = TRUE),
+  pattern_extracted <- factor(terra::extract(x = raster, y = pattern, factors = FALSE)[, 2],
                               levels = habitat_levels)
 
   result <- utils::stack(table(pattern_extracted)) # count number of points within each habitat