+++ R/print.rd_ras.R
@@ -17,7 +17,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' landscape_classified <- classify_habitats(landscape, n = 5, style = "fisher")
+#' landscape_classified <- classify_habitats(terra::rast(landscape), n = 5, style = "fisher")
 #' landscape_random <- randomize_raster(landscape_classified, n_random = 19)
 #'
 #' print(landscape_random)
@@ -27,21 +27,17 @@
 #' @rdname print.rd_ras
 #'
 #' @export
-print.rd_ras <- function(x,
-                         ...) {
+print.rd_ras <- function(x, ...) {
 
-  # set length observed pattern to 0 and
-  # return warning that energy can't be calculated
-  if (!methods::is(x$observed, "RasterLayer")) {
+  # check if observed raster is included
+  if (!methods::is(x$observed, "SpatRaster")) {
 
     number_raster_obs <- 0
 
     includes_observed <- "NA"
 
-  }
-
   # observed pattern is present
-  else {
+  } else {
 
     number_raster_obs <- 1
 
@@ -50,10 +46,8 @@ print.rd_ras <- function(x,
   }
 
   # get extent of window
-  extent_window <- paste0(c(raster::xmin(x$randomized[[1]]),
-                            raster::xmax(x$randomized[[1]]),
-                            raster::ymin(x$randomized[[1]]),
-                            raster::ymax(x$randomized[[1]])),
+  extent_window <- paste0(c(terra::xmin(x$randomized[[1]]), terra::xmax(x$randomized[[1]]),
+                            terra::ymin(x$randomized[[1]]), terra::ymax(x$randomized[[1]])),
                           collapse = " ")
 
   # get number of randomized patterns plus observed pattern