+++ data-raw/example_data.R
@@ -1,8 +1,9 @@
 library(dplyr)
-library(NLMR)
-library(maptools)
+# library(NLMR)
 library(usethis)
+library(sf)
 library(spatstat)
+library(terra)
 
 library(shar)
 
@@ -10,22 +11,24 @@ library(shar)
 
 set.seed(42)
 
-# Create landscape
-landscape <- NLMR::nlm_fbm(ncol = 50, nrow = 50, resolution = 20,
-                           fract_dim = 1.5, user_seed = 42)
+# # Create landscape
+# landscape <- NLMR::nlm_fbm(ncol = 50, nrow = 50, resolution = 20,
+#                            fract_dim = 1.5, user_seed = 42) %>%
+#   terra::rast()
+
+landscape <- terra::rast(ncol = 50, nrow = 50, resolution = 20, extent = c(0, 1000, 0, 1000),
+                         crs = NA, vals = seq(from = 0, to = 1, length.out = 2500))
 
 landscape_class <- classify_habitats(landscape, n = 5, style = "fisher")
 
 # Create species with negative
-pattern_a <- spatstat.random::runifpoint(n = 250, win = spatstat.geom::owin(c(0, 1000),
-                                                                          c(0, 1000)))
+pattern_a <- spatstat.random::runifpoint(n = 250, win = spatstat.geom::owin(c(0, 1000), c(0, 1000)))
 
 # get habitat 4 as owin
-owin_pattern <- raster::rasterToPolygons(landscape_class,
-                                         fun = function(x){x == 4},
-                                         dissolve = TRUE) %>%
-  maptools::as.owin.SpatialPolygons()
-
+owin_pattern <- terra::as.polygons(landscape_class) %>%
+  terra::subset(.$lyr.1 == 4) %>%
+  sf::st_as_sf() %>%
+  spatstat.geom::as.owin()
 
 # create pattern with no point in habitat 4
 species_a <- pattern_a[!spatstat.geom::inside.owin(x = pattern_a, w = owin_pattern)]
@@ -43,13 +46,14 @@ marks_df_a <- data.frame(status = factor(sample(c("dead", "alive"),
 spatstat.geom::marks(species_a) <- marks_df_a
 
 # Create species with positive associations
-pattern_b <- spatstat.random::runifpoint(n = species_a$n / 2,
-                                       win = spatstat.geom::owin(c(0, 1000), c(0, 1000)))
+pattern_b <- spatstat.random::runifpoint(n = floor(species_a$n / 2),
+                                         win = spatstat.geom::owin(c(0, 1000), c(0, 1000)))
 
 # create pattern with more points in habitat 5
-species_b <- raster::rasterToPolygons(landscape_class,
-                                      fun = function(x){x == 5}, dissolve = TRUE) %>%
-  maptools::as.owin.SpatialPolygons() %>%
+species_b <- terra::as.polygons(landscape_class) %>%
+  terra::subset(.$lyr.1 == 5) %>%
+  sf::st_as_sf() %>%
+  spatstat.geom::as.owin() %>%
   spatstat.random::runifpoint(n = floor(pattern_b$n * 1), win = .) %>%
   spatstat.geom::superimpose.ppp(pattern_b, W = spatstat.geom::owin(c(0, 1000), c(0, 1000)))
 
@@ -59,7 +63,7 @@ marks_df_b <- factor(sample(c("dominant", "understorey"),
 spatstat.geom::marks(species_b) <- marks_df_b
 
 # set number of repetitions
-n_random <- 99
+n_random <- 3
 
 # translate raster
 torus_trans <- translate_raster(raster = landscape_class)
@@ -76,9 +80,10 @@ reconstruction <- reconstruct_pattern(pattern = species_b, n_random = n_random,
 
 #### Save data ####
 
-overwrite <- FALSE
+overwrite <- TRUE
 
 # save landscape
+landscape <- terra::wrap(landscape)
 usethis::use_data(landscape, overwrite = overwrite)
 
 # save species data
@@ -87,8 +92,10 @@ usethis::use_data(species_a, overwrite = overwrite)
 usethis::use_data(species_b, overwrite = overwrite)
 
 # save random landscape data
+torus_trans <- pack_randomized(raster = torus_trans)
 usethis::use_data(torus_trans, overwrite = overwrite)
 
+random_walk <- pack_randomized(raster = random_walk)
 usethis::use_data(random_walk, overwrite = overwrite)
 
 # save random point data